backend/requirements.txt
flask==3.0.3
flask-cors==4.0.1
python-dotenv==1.0.1
supabase==2.4.6
google-generativeai==0.7.2
